move links under entry
