+++ core/src/action.rs
@@ -1,7 +1,9 @@
 use agent::state::AgentState;
 use context::Context;
 use holochain_core_types::{
-    cas::content::Address, entry::Entry, error::HolochainError, links_entry::Link,
+    cas::content::Address,
+    entry::{link_add::Link, Entry},
+    error::HolochainError,
     validation::ValidationPackage,
 };
 use holochain_dna::Dna;
