+++ core/src/dht/dht_reducers.rs
@@ -219,8 +219,7 @@ pub mod tests {
     };
     use holochain_core_types::{
         cas::content::AddressableContent,
-        entry::{test_entry, test_sys_entry, test_unpublishable_entry, Entry},
-        links_entry::Link,
+        entry::{link_add::Link, test_entry, test_sys_entry, test_unpublishable_entry, Entry},
     };
     use instance::tests::test_context;
     use state::test_store;
