+++ core/src/dht/dht_store.rs
@@ -5,9 +5,9 @@ use holochain_core_types::{
         storage::ContentAddressableStorage,
     },
     eav::{EntityAttributeValue, EntityAttributeValueStorage},
+    entry::link_add::Link,
     error::HolochainError,
     hash::HashString,
-    links_entry::Link,
 };
 use std::{
     collections::{HashMap, HashSet},
