+++ core_types/src/cas/storage.rs
@@ -3,6 +3,7 @@ use eav::{EntityAttributeValue, EntityAttributeValueStorage};
 use entry::{test_entry_unique, SerializedEntry};
 use error::HolochainError;
 use json::RawString;
+use objekt;
 use std::{
     collections::{HashMap, HashSet},
     fmt::Debug,
