+++ core_types/src/entry/link_add.rs
@@ -1,13 +1,13 @@
+//-------------------------------------------------------------------------------------------------
+// Link
+//-------------------------------------------------------------------------------------------------
+
 use cas::content::Address;
 use entry::{Entry, ToEntry};
 use entry_type::EntryType;
 use error::HolochainError;
 use json::JsonString;
 use std::convert::TryInto;
-
-//-------------------------------------------------------------------------------------------------
-// Link
-//-------------------------------------------------------------------------------------------------
 
 type LinkTag = String;
 
@@ -98,51 +98,16 @@ impl ToEntry for LinkEntry {
     }
 }
 
-//-------------------------------------------------------------------------------------------------
-// LinkListEntry
-//-------------------------------------------------------------------------------------------------
-//
-#[derive(Serialize, Deserialize, PartialEq, Clone, Debug, DefaultJson)]
-pub struct LinkListEntry {
-    links: Vec<Link>,
-}
-
-impl LinkListEntry {
-    pub fn new(links: &[Link]) -> Self {
-        LinkListEntry {
-            links: links.to_vec(),
-        }
-    }
-
-    pub fn links(&self) -> &Vec<Link> {
-        &self.links
-    }
-}
-
-impl ToEntry for LinkListEntry {
-    // Convert a LinkListEntry into a JSON array of Links
-    fn to_entry(&self) -> Entry {
-        Entry::new(EntryType::LinkList, self.to_owned())
-    }
-
-    fn from_entry(entry: &Entry) -> Self {
-        assert_eq!(&EntryType::LinkList, entry.entry_type());
-        entry
-            .value()
-            .to_owned()
-            .try_into()
-            .expect("could not convert Entry to LinkListEntry")
-    }
-}
-
 #[cfg(test)]
 pub mod tests {
 
     use cas::content::AddressableContent;
-    use entry::{test_entry_a, test_entry_b, Entry, ToEntry};
+    use entry::{
+        link_add::{Link, LinkActionKind, LinkEntry, LinkTag},
+        test_entry_a, test_entry_b, Entry, ToEntry,
+    };
     use entry_type::EntryType;
     use json::JsonString;
-    use links_entry::{Link, LinkActionKind, LinkEntry, LinkTag};
     use std::convert::TryFrom;
 
     pub fn test_link_tag() -> LinkTag {
