+++ core_types/src/entry/mod.rs
@@ -11,6 +11,8 @@ use std::{
     ops::Deref,
 };
 pub mod agent;
+pub mod link_add;
+pub mod link_list;
 
 pub type EntryValue = JsonString;
