+++ core_types/src/lib.rs
@@ -37,7 +37,6 @@ pub mod file_validation;
 pub mod hash;
 pub mod json;
 pub mod keys;
-pub mod links_entry;
 pub mod signature;
 pub mod time;
 pub mod validation;
